+++ js/constants.js
@@ -1,0 +1,2 @@
+export const W = 960;
+export const H = 600;